Projet 4 - Détection de Fraude dans les Transactions Bancaires/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyspark>=3.5.0
google-cloud-bigquery>=3.14.0
google-cloud-bigquery-storage>=2.24.0
pyarrow>=14.0.0
apache-airflow>=2.8.0
streamlit>=1.30.0
plotly>=5.18.0
scikit-learn>=1.3.0
db-dtypes>=1.2.0
python-dateutil>=2.8.0
